✨ FEATURE : Code for generating double-moon shape datasets
This function takes seven inputs: three optional and four mandatory. The radius, width, rotation, and separation distance are required arguments. The size of the dataset, the proportion of the train data to the whole dataset, and one flag to plot the dataset are the other inputs.

In the end, its outputs are the train and test datasets. These matrices contain the labels of the inputs, such as +1 and -1, in addition to the inputs x1 and x2.

Consequently, different neural networks might be trained using these datasets.